+++ test/helpers/mock_annotations.mocks.dart
@@ -1930,7 +1930,7 @@ class MockCatalogueRemoteDataSource extends _i1.Mock
   _i5.Future<void> updateStock(
     String? accountId,
     String? productId,
-    int? newStock,
+    double? newStock,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
